Fixed slow pad generation issue
So here's the deal: one-time pad generation was slow for Linux systems,
but not macOS systems, before now. This was due to the use of
`/dev/random` to generate the pads, which works differently on different
these systems. `/dev/random` attempts to understand the entropy of the
data it outputs and blocks until entropy is high enough, but this
results in very slow read times and long periods of blocking as it waits
for more entropy to enter the system. Even with this, `/dev/random`
isn't a perfect source of random numbers. For more on this, see:
https://www.2uo.de/myths-about-urandom/

Due to this issue, I've switched from reading from `dev/random` to using
the system-provided `getentropy()` available on both macOS and Linux
systems. This avoids the primary pitfall with using `/dev/urandom`,
which doesn't block, by ensuring that there's enough entropy when random
number generation starts to produce reliably unpredictable data.